fix(worker): 清理死字段、提取公共辅助函数、时间戳改毫秒
- 删除 bundle_name/ability_name 死字段：harmony.py 实际使用
  action.value/action.name，action.py 中的字段定义、from_dict
  解析、to_dict 序列化及对应测试改为基于 value/name
- 提取 _get_timestamp/_image_to_base64/_region_b64/_parse_row_tolerance
  到 worker/actions/base.py，image.py/ocr.py/coordinate.py 改为 import
- _get_timestamp 返回改为毫秒（int(time.time()*1000)），
  ActionResult.timestamp 类型 float→int，注释同步
- row_tolerance 类型注解精确化为 int | list[int] | None

+++ worker/task/result.py
@@ -52,7 +52,7 @@ class ActionResult:
     # 区域截图（same_row 等使用裁剪区域时，附带裁剪后的小图）
     region_screenshot: Optional[str] = None  # base64，失败时的裁剪区域截图
     # 点击类动作时间戳
-    timestamp: Optional[float] = None  # 点击动作执行时的时间戳（Unix timestamp）
+    timestamp: Optional[int] = None  # 点击动作执行时的时间戳（毫秒级）
     artifacts: List[Dict[str, Any]] = field(default_factory=list)
 
     @classmethod
